Move mock to its own package
