+++ src/go/userepository/list_repositories_query_test.go
@@ -1,6 +1,7 @@
 package userepository_test
 
 import (
+	repomock "github.com/kkrull/marmot/corerepositorymock"
 	main "github.com/kkrull/marmot/mainfactory"
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
@@ -9,12 +10,12 @@ import (
 var _ = Describe("ListRepositoriesQuery", func() {
 	var (
 		factory *main.CommandFactory
-		source  *MockRepositorySource
+		source  *repomock.RepositorySource
 	)
 
 	Describe("#Run", func() {
 		It("returns all repositories the source can list", func() {
-			source = &MockRepositorySource{Names: []string{"one", "two"}}
+			source = &repomock.RepositorySource{Names: []string{"one", "two"}}
 			factory = &main.CommandFactory{RepositorySource: source}
 
 			subject, factoryErr := factory.ListRepositoriesQuery()
